LayoutTests/media/track/track-webvtt-tc011-blanklines-expected.txt

Tests that cues are not affected by multiple newlines \n, \r, and \r\n and that cue not properly separated are treated as one big cue.

EXPECTED (video.textTracks[0].cues.length == '3') OK
EXPECTED (video.textTracks[0].cues[0].id == '1') OK
EXPECTED (video.textTracks[0].cues[0].startTime == '0.0') OK
EXPECTED (video.textTracks[0].cues[0].endTime == '30.5') OK
EXPECTED (video.textTracks[0].cues[0].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (video.textTracks[0].cues[1].id == '2') OK
EXPECTED (video.textTracks[0].cues[1].startTime == '31.0') OK
EXPECTED (video.textTracks[0].cues[1].endTime == '60.5') OK
EXPECTED (video.textTracks[0].cues[1].getCueAsSource() == 'I said Bear is coming!!!!') OK
EXPECTED (video.textTracks[0].cues[2].id == '3') OK
EXPECTED (video.textTracks[0].cues[2].startTime == '61.0') OK
EXPECTED (video.textTracks[0].cues[2].endTime == '361200.5') OK
EXPECTED (video.textTracks[0].cues[2].getCueAsSource() == 'I said Bear is coming now!!!!') OK
EXPECTED (video.textTracks[1].cues.length == '1') OK
END OF TEST
